+++ tests/unit_test/moss_tts/test_pipeline.py
@@ -673,6 +673,7 @@ def test_moss_tts_preprocessing_uses_placement_gpu_id(
     )
     encoder = SimpleNamespace()
     loaded: list[tuple[str, str, torch.dtype | None]] = []
+    resolved: list[tuple[str | None, int | None]] = []
 
     def load_encoder(
         model_path,
@@ -687,6 +688,11 @@ def test_moss_tts_preprocessing_uses_placement_gpu_id(
 
     monkeypatch.setattr(stages, "_load_moss_processor", lambda model_path: processor)
     monkeypatch.setattr(stages, "load_moss_audio_encoder", load_encoder)
+    monkeypatch.setattr(
+        stages,
+        "resolve_device_spec",
+        lambda device, gpu_id: resolved.append((device, gpu_id)) or "resolved:2",
+    )
 
     try:
         stages.create_preprocessing_executor(
@@ -700,12 +706,8 @@ def test_moss_tts_preprocessing_uses_placement_gpu_id(
     finally:
         rb.clear_moss_tts_preprocessing_context()
 
-    # The contract is that placement's gpu_id reaches the codec, not that the
-    # device is CUDA — the stage resolves the type from the live platform, so a
-    # pinned "cuda:2" here would only assert the host this test happened to run on.
-    from sglang_omni.utils.device import resolve_device_spec
-
-    assert loaded == [("codec", resolve_device_spec(None, 2), torch.bfloat16)]
+    assert resolved == [(None, 2)]
+    assert loaded == [("codec", "resolved:2", torch.bfloat16)]
 
 
 def test_moss_tts_pathlike_reference_uses_separate_codec() -> None:
